test: improve ut coverage
fix
